Update docs for metrics
