+++ torchbearer/metrics/roc_auc_score.py
@@ -13,19 +13,16 @@ class RocAucScore(metrics.EpochLambda):
 
         Requires :mod:`sklearn.metrics`.
 
+    :param one_hot_labels: If True, convert the labels to a one hot encoding. Required if they are not already.
+    :type one_hot_labels: bool
+    :param one_hot_offset: Subtracted from class labels, use if not already zero based.
+    :type one_hot_offset: int
+    :param one_hot_classes: Number of classes for the one hot encoding.
+    :type one_hot_classes: int
+
     """
 
     def __init__(self, one_hot_labels=True, one_hot_offset=0, one_hot_classes=10):
-        """Create the metric with the given configuration.
-
-        :param one_hot_labels: If True, convert the labels to a one hot encoding. Required if they are not already.
-        :type one_hot_labels: bool
-        :param one_hot_offset: Subtracted from class labels, use if not already zero based.
-        :type one_hot_offset: int
-        :param one_hot_classes: Number of classes for the one hot encoding.
-        :type one_hot_classes: int
-
-        """
 
         def to_categorical(y):
             return np.eye(one_hot_classes, dtype='uint8')[y - one_hot_offset]
